Invert integer weight to float

+++ src/main/java/com.goldpoisk.app/Gem.java
@@ -10,7 +10,7 @@ public class Gem {
     String name;
     Float weight;
 
-    public Gem(String name, Integer weight) {
+    public Gem(String name, Float weight) {
         this.name = name;
         this.weight = weight;
     }
